Fix: 프리즈마, 회원가입 로그인기능 오류나는 부분 수정 및 라이브러리 설치 완료
