+++ prisma/schema.prisma
@@ -1,0 +1,16 @@
+generator client {
+ provider = "prisma-client-js"
+}
+datasource db {
+ provider = "mysql"
+ url   = env("DATABASE_URL")
+}
+model Users {
+ userId    Int      @id @default(autoincrement())
+ isClient  Boolean
+ email     String   @unique
+ password  String
+ name      String
+ createdAt DateTime @default(now())
+ updatedAt DateTime @updatedAt
+}
