--- src/auth/dtos/login.dto.ts
@@ -1,16 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsEmail, IsNotEmpty, IsString, MinLength } from 'class-validator';
-
-export class LoginDto {
-    @IsEmail()
-    @IsNotEmpty()
-    @ApiProperty()
-    email: string;
-  
-    @IsString()
-    @IsNotEmpty()
-    @MinLength(4)
-    @ApiProperty()
-    password: string;
-  }
-
