+++ src/users/users.controller.ts
@@ -1,9 +1,9 @@
-import { Controller, Post } from '@nestjs/common';
+import { Controller } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { UsersService } from './users.service';
 
 @Controller('users')
 @ApiTags('users')
 export class UsersController {
-    constructor(private readonly usersService:UsersService){}
+  constructor(private readonly usersService: UsersService) {}
 }
